pyproject.toml
[project]
name = "contract-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pact-python>=3.2.1",
    "pact-python-cli>=2.5.7.0",
    "psutil>=7.2.1",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "six>=1.17.0",
    "uvicorn>=0.40.0",
]
